+++ internal/commands/util.go
@@ -7,7 +7,6 @@ import (
 	"strings"
 
 	"github.com/UpCloudLtd/upcloud-go-api/v4/upcloud"
-	"github.com/jedib0t/go-pretty/v6/text"
 
 	"github.com/UpCloudLtd/upcloud-cli/internal/validation"
 )
@@ -68,46 +67,6 @@ func SearchResources(
 		}
 	}
 	return result, nil
-}
-
-// DatabaseStateColour maps database states to colours
-func DatabaseStateColour(state upcloud.ManagedDatabaseState) text.Colors {
-	switch state {
-	case upcloud.ManagedDatabaseStateRunning:
-		return text.Colors{text.FgGreen}
-	case "rebuilding", "rebalancing":
-		return text.Colors{text.FgYellow}
-	default:
-		return text.Colors{text.FgHiBlack}
-	}
-}
-
-// LoadBalancerOperationalStateColour maps load balancer states to colours
-func LoadBalancerOperationalStateColour(state upcloud.LoadBalancerOperationalState) text.Colors {
-	switch state {
-	case upcloud.LoadBalancerOperationalStateRunning:
-		return text.Colors{text.FgGreen}
-	case upcloud.LoadBalancerOperationalStateCheckup, upcloud.LoadBalancerOperationalStatePending, upcloud.LoadBalancerOperationalStateSetupAgent, upcloud.LoadBalancerOperationalStateSetupDNS, upcloud.LoadBalancerOperationalStateSetupLB, upcloud.LoadBalancerOperationalStateSetupNetwork, upcloud.LoadBalancerOperationalStateSetupServer:
-		return text.Colors{text.FgYellow}
-	default:
-		return text.Colors{text.FgHiBlack}
-	}
-}
-
-// StorageStateColour is a helper mapping storage states to colours
-func StorageStateColour(state string) text.Colors {
-	switch state {
-	case upcloud.StorageStateOnline, upcloud.StorageStateSyncing:
-		return text.Colors{text.FgGreen}
-	case upcloud.StorageStateError:
-		return text.Colors{text.FgHiRed, text.Bold}
-	case upcloud.StorageStateMaintenance:
-		return text.Colors{text.FgYellow}
-	case upcloud.StorageStateCloning, upcloud.StorageStateBackuping:
-		return text.Colors{text.FgHiMagenta, text.Bold}
-	default:
-		return text.Colors{text.FgHiBlack}
-	}
 }
 
 // BoolFromString parses a string and returns *upcloud.Boolean