+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- In `storage list` output: capitalize zone column header and color storage state similarly than in `storage show`.
 
 ## [2.0.0] - 2022-08-30
 ### Added